#22 Fixed TypeError 'integrity' of undefined.
In some instances, the src is not found from the compilation.assets array causing a build error. This occures at least with css files.
This fix allows to skip these items that produces a correct warning of:  Cannot determine hash for asset 'css/app.xxxxxxx.css', the resource will be unprotected.