Revert "Merge pull request #85 from waysact/safe-require"
On second thought the change in #85 introduces the first non-essential
run-time dependency which, for a small piece of code like this,
doesn't seem to be worth it.  I'd rather keep run-time dependencies
limited to the essentials, especially when the extra dependency
doesn't save more than a couple of lines of code.

This reverts commit ad1bfab96efb615cd3800d0814653ad22a8c84ac, reversing
changes made to 488459a043a6d9347222168306b8ab55c40c4743.